Update system-wide permission add/remove
Also remove refresh user permission ability.
- This is a emergency-only kind of admin function so we may add it back later

+++ src/Controllers/Permissions/RemovePermissionsController.php
@@ -7,36 +7,36 @@
 
 namespace Hal\UI\Controllers\Permissions;
 
-use Hal\Core\Entity\UserPermission;
+use Doctrine\ORM\EntityManagerInterface;
+use Hal\Core\Entity\User;
+use Hal\Core\Entity\User\UserPermission;
+use Hal\UI\Controllers\CSRFTrait;
 use Hal\UI\Controllers\RedirectableControllerTrait;
 use Hal\UI\Controllers\SessionTrait;
-use Hal\UI\Flash;
+use Hal\UI\Controllers\TemplatedControllerTrait;
+use Hal\UI\Security\AuthorizationHydrator;
 use Hal\UI\Security\AuthorizationService;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use QL\Panthor\ControllerInterface;
+use QL\Panthor\TemplateInterface;
 use QL\Panthor\Utility\URI;
 
-/**
- * Super:
- *     Add any.
- *     Remove Lead, ButtonPusher
- *
- * ButtonPusher:
- *     Add Lead, ButtonPusher
- *     Remove Lead
- */
 class RemovePermissionsController implements ControllerInterface
 {
+    use CSRFTrait;
     use RedirectableControllerTrait;
+    use RemovePermissionsTrait;
     use SessionTrait;
+    use TemplatedControllerTrait;
 
-    const MSG_SUCCESS = 'User Permission "%s" revoked from "%s".';
+    private const MSG_SUCCESS = 'User Permission "%s" revoked from "%s".';
+    private const ERR_DENIED = 'Access Denied';
 
     /**
-     * @var AuthorizationService
+     * @var TemplateInterface
      */
-    private $authorizationService;
+    private $template;
 
     /**
      * @var URI
@@ -44,13 +44,37 @@ class RemovePermissionsController implements ControllerInterface
     private $uri;
 
     /**
+     * @var AuthorizationService
+     */
+    private $authorizationService;
+
+    /**
+     * @var AuthorizationHydrator
+     */
+    private $authorizationHydrator;
+
+    /**
+     * @param TemplateInterface $template
+     * @param EntityManagerInterface $em
      * @param AuthorizationService $authorizationService
+     * @param AuthorizationHydrator $authorizationHydrator
      * @param URI $uri
      */
-    public function __construct(AuthorizationService $authorizationService, URI $uri)
-    {
+    public function __construct(
+        TemplateInterface $template,
+        EntityManagerInterface $em,
+        AuthorizationService $authorizationService,
+        AuthorizationHydrator $authorizationHydrator,
+        URI $uri
+    ) {
+        $this->template = $template;
+
         $this->authorizationService = $authorizationService;
+        $this->authorizationHydrator = $authorizationHydrator;
+
         $this->uri = $uri;
+
+        $this->setEntityManagerForRemovalPermissions($em);
     }
 
     /**
@@ -58,15 +82,53 @@ class RemovePermissionsController implements ControllerInterface
      */
     public function __invoke(ServerRequestInterface $request, ResponseInterface $response)
     {
+        $currentUserAuthorizations = $this->getAuthorizations($request);
+
+        $user = $request->getAttribute(User::class);
         $permission = $request->getAttribute(UserPermission::class);
 
-        $type = $permission->type();
-        $name = $permission->user()->username();
+        if (!$this->isRemovalAllowed($currentUserAuthorizations, $permission)) {
+            $this->withFlashError($request, self::ERR_DENIED, $this->getRemovalDeniedReason());
+            return $this->withRedirectRoute($response, $this->uri, 'admin.permissions');
+        }
+
+        if ($this->handleForm($request, $permission)) {
+            $type = $permission->type();
+            $name = $permission->user()->name();
+
+            $this->withFlashSuccess($request, sprintf(self::MSG_SUCCESS, $type, $name));
+            return $this->withRedirectRoute($response, $this->uri, 'admin.permissions');
+        }
+
+        $selectedUserAuthorizations = $this->authorizationService->getUserAuthorizations($user);
+        $selectedUserPerms = $this->authorizationHydrator->hydrateAuthorizations($user, $selectedUserAuthorizations);
+
+        return $this->withTemplate($request, $response, $this->template, [
+            'user' => $user,
+            'user_authorizations' => $selectedUserAuthorizations,
+            'user_permissions' => $selectedUserPerms,
+
+            'permission' => $permission
+        ]);
+    }
+
+    /**
+     * @param ServerRequestInterface $request
+     * @param UserPermission $permission
+     *
+     * @return bool
+     */
+    private function handleForm(ServerRequestInterface $request, UserPermission $permission)
+    {
+        if ($request->getMethod() !== 'POST') {
+            return false;
+        }
+
+        if (!$this->isCSRFValid($request)) {
+            return false;
+        }
 
         $this->authorizationService->removeUserPermissions($permission);
-
-        $this->withFlash($request, Flash::SUCCESS, sprintf(self::MSG_SUCCESS, $type, $name));
-
-        return $this->withRedirectRoute($response, $this->uri, 'admin.permissions');
+        return true;
     }
 }
